+++ src/main/java/arez/spytools/browser/package-info.java
@@ -1,7 +1,4 @@
 /**
  * Browser based introspection tools based on the core Arez spy infrastructure.
  */
-@Unsupported( "These utilities are based on unsupported core.spy package and are expected to evolve as needs dictate" )
 package arez.spytools.browser;
-
-import org.realityforge.anodoc.Unsupported;