+++ src/main/java/arez/spytools/package-info.java
@@ -1,7 +1,4 @@
 /**
  * Additional utilities and introspection tools based on the core Arez spy infrastructure.
  */
-@Unsupported( "These utilities are based on unsupported arez.spy package and are expected to evolve as needs dictate" )
 package arez.spytools;
-
-import org.realityforge.anodoc.Unsupported;