+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Remove usage of the `org.realityforge.anodoc` dependency.
+
 ### [v0.15](https://github.com/arez/arez-spytools/tree/v0.15) (2018-06-13)
 [Full Changelog](https://github.com/arez/arez-spytools/compare/v0.14...v0.15)